requirements.txt
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.0
jsonschema==3.2.0
numpy==1.18.4
nyoka==4.2.0
olefile==0.46
packaging==19.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.10
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
Pygments==2.4.2
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
pyparsing==2.4.5
pyreadline==2.1
pyrsistent==0.15.5
pytest==5.0.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==223
pywinpty==0.5.5
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.4.1
sklearn-pandas==1.8.0
sklearn2pmml==0.56.2
statsmodels==0.10.1

